os/windows/waitable: Workaround for a weird issue with async io (part 3)
For an unknown reason, ReadFileEx and WriteFileEx will not correctly set the Overlapped Event and instead only wake up any currently waiting threads. This means that in the event that a race condition occurs and the event completed before you were able to wait for it, you were stuck or timing out repeatedly. This fixes the issue by manually setting the event into the correct state, so that WaitForSingleObjectEx will return WAIT_OBJECT_0 instead of WAIT_TIMEOUT.