os/windows/named-pipe: Update to async_request::completion_routine
This adds support for callbacks in async_op, which are run whenever the waitable read()/write() is signalled. Callbacks do not work for other requests, like accept().

Generally callbacks should be used sparingly as they might not be supported across platforms. They are only there to ease the coding effort on various platforms that do support them.