os/async_op: Add proper callback routines
These callbacks are to be used on completion of something, so on windows these replace the completion routine that was used before. This improves speed a little bit as the kernel can resume operation quicker.

These callbacks can only be called once and then the async_op needs to be reset.