os/windows/named-pipe: Fix is_connected and remove deprecated methods
The `is_connected` method will now return false if the processId and sessionId no longer match the known one through accept or open. In this case, the socket is considered disconnected, and created sockets have to use accept to do anything again.

Additionally several legacy methods were removed so that current code does not have to deal with legacy behavior.